added meta data to contactForm, policy, terms, cookie and small text fixes
