+++ messages/ar.json
@@ -131,11 +131,11 @@
       },
       {
         "title": "برامج متخصّصة لكل قطاع",
-        "subtitle": "يتماشى مع أدوات مؤسستك، وأنظمتها، ومعايير الامتثال في قطاعك"
+        "subtitle": "تتماشى مع أدوات مؤسستك، وأنظمتها، ومعايير الامتثال في قطاعك"
       },
       {
         "title": "أكاديميّة رقميّة بهويّة مؤسّستك",
-        "subtitle": "اكاديمية رقمية ليست منصة وحسب. دورات تفاعلية في شكل رحلات تدريبية مع نخبة من الخبراء. فريق دعم ومحتوى متنوع"
+        "subtitle": "أكاديمية رقمية ليست منصة وحسب. دورات تفاعلية في شكل رحلات تدريبية مع نخبة من الخبراء. فريق دعم ومحتوى متنوع"
       },
       {
         "title": "حضوري. عن بُعد. ذاتي. مدمج",
@@ -147,7 +147,7 @@
       }
     ],
     "title": "حلول تدريبية مصمّمة خصّيصًا لمؤسستك  ",
-    "bottomTitle": "هويته. رؤيتها. أهدافها",
+    "bottomTitle": "هويتها. رؤيتها. أهدافها",
     "subtitle": "احتياج مؤسستك الفعلي نقطة البداية التي نصمّم الحلول حولها"
   },
   "FutureSection": {
@@ -442,9 +442,9 @@
     ],
     "additional_info": "معلومات إضافية",
     "additional_info_placeholder": "هل هناك تفاصيل أو أهداف محددة ترغب بمشاركتها؟",
-    "agreement_1": "أوافق على",
+    "agreement_1": "أوافق على ",
     "agreement_terms": "الشروط",
-    "agreement_and": " و",
+    "agreement_and": " و ",
     "agreement_privacy": "الأحكام",
     "submit": "احجز عرض توضيحي"
   },
