+++ src/app/[locale]/cookie-policy/page.tsx
@@ -1,11 +1,25 @@
+import type { Metadata } from 'next'
 import Footer from '@/components/Footer/Footer'
 import { Header } from '@/components/Header'
-import { Metadata } from 'next'
 import { useTranslations } from 'next-intl'
 
-export const metadata: Metadata = {
-  title: 'Cookie Policy | Coursinity',
-  description: 'Cookie policy for Coursinity services and platform.'
+export async function generateMetadata({ params }: { params: Promise<{ locale: string }> }): Promise<Metadata> {
+  const { locale } = await params
+  const isArabic = locale === 'ar'
+
+  return {
+    title: isArabic
+      ? 'أفضل شركات التدريب في السعودية | Coursinity'
+      : 'Best Training Companies in Saudi Arabia | Coursinity',
+    description: isArabic
+      ? 'الرائدة في رفع إنتاجية أعضاء فريقك بأحدث البرامج والتقنيات'
+      : 'Explore COURSINITY: Blended courses that combine real-life experience and interactive learning, for the best professional certifications and internationally accredited training in Saudi Arabia.',
+    icons: {
+      icon: '/assets/favicon.png',
+      shortcut: '/assets/favicon.png',
+      apple: '/assets/favicon.png'
+    }
+  }
 }
 
 export default function CookiePolicyPage() {
